refactor(io): lift style and column decoding out of the sheet data reader (spec 28 task 6)
WorksheetSheetDataReader is back to reading sheet data. Of the ten members that had nothing
to do with it, four went in task 3 (LoadStyleFill, LoadStyleBorder, LoadStyleFont,
LoadStyleNumberFormat), UInt32HasValue went with them, the phonetics font decode went in
task 4, and the last five move here:

  ApplyStyle, ResolveStyleValue -> StyleDecoder
  LoadStyle                     -> StyleDecoder as a Decode(int styleIndex, ...) overload
  LoadColumns, LoadColumn       -> new WorksheetColumnReader

LoadStyle had shrunk to a guard plus one call in task 3. Rather than inline it into its
three callers and copy the "no stylesheet, no styles" guard three times, it becomes a third
Decode overload alongside Decode(CellFormat, ...) and Decode(DifferentialFormat, ...), and
the ref-parameter signature goes away. XLWorkbook_Load's normal-style call site now reads
as an assignment rather than a mutation.

<cols> is a worksheet element in its own right, not sheet data, and the loader already
dispatches it as one - spec 24 landed, so the call site moved is WorksheetElementReader's,
not XLWorkbook_Load's.

The class summary no longer claims to do "style application".

ONE ACCEPTANCE GATE DOES NOT RETURN 0, AND THAT IS THE SPEC CONTRADICTING ITSELF

Criterion 6 asks that
  grep -cE '(LoadStyle|ApplyStyle|ResolveStyleValue|UInt32HasValue|LoadColumn)' \
    XLibur/Excel/IO/WorksheetSheetDataReader.cs
return 0. It returns 1: line 1003, `StyleDecoder.ApplyStyle(xlRow, ...)` inside
ApplyRowCustomProps - a qualified CALL, not a declaration.

That call cannot go without duplicating ApplyStyle's body at the call site, and task 6 step
4 of the same spec lists ApplyRowCustomProps among the matches it EXPECTS to remain, on the
grounds that it is a row using a resolved style rather than decoding one. The prose of
criterion 6 - "WorksheetSheetDataReader.cs declares none of ..." - is satisfied: every one
of the ten is gone as a declaration. Copying a method body to make a grep count zero would
be the wrong trade, so the call stays and the discrepancy is reported instead.

Diff shape: WorksheetSheetDataReader -93/+7, now 1272 lines from 1460.

Suite: both target frameworks, 28292 total, 0 failed, 28282 passed, 10 skipped.

Spec: 28-single-style-decoder.md task 6.
