+++ XLibur/Excel/XLWorkbook_Load.cs
@@ -145,8 +145,8 @@ public partial class XLWorkbook
             .FirstOrDefault(x => x.BuiltinId is not null && x.BuiltinId.Value == 0);
         if (normalStyle != null)
         {
-            var normalStyleKey = ((XLStyle)Style).Key;
-            WorksheetSheetDataReader.LoadStyle(ref normalStyleKey, (int)normalStyle.FormatId!.Value, context.Styles);
+            var normalStyleKey = StyleDecoder.Decode((int)normalStyle.FormatId!.Value,
+                context.Styles, ((XLStyle)Style).Key);
             Style = new XLStyle(null!, normalStyleKey);
             ColumnWidth = CalculateColumnWidth(8, Style.Font, this);
         }
@@ -310,7 +310,7 @@ public partial class XLWorkbook
                 continue;
             }
 
-            WorksheetSheetDataReader.ApplyStyle(ws, 0, styles);
+            StyleDecoder.ApplyStyle(ws, 0, styles);
 
             LoadWorksheetElements(worksheetPart, ws, sharedStrings, context);
